Refactor: move entrypoint from __init__
