+++ tests/test_main.py
@@ -1,0 +1,10 @@
+from traceback import print_tb
+
+from elisctl.main import entry_point
+
+
+class TestEntryPoint:
+    def test_shell(self, cli_runner):
+        result = cli_runner.invoke(entry_point)
+        assert not result.exit_code, print_tb(result.exc_info[2])
+        assert "elis> " == result.output
